Add quickcheck tests around the simulated syncing process.
This actually found more bugs in memory_replica than in the actual
syncing process (only 1 for the latter).

Unfortunately most of the changes are log plumbing dealing with the
bizarre non-composable design and inaccessible API of rust's test output
capturing. It's also a bit unfortunate that running the tests even in
release mode now takes 2m (and 4m in debug mode).