Add new storage mechanism for discovering dirty server directories.
`for_dir` has a number of problems. The biggest one comes into play if
multiple server roots are in use, as every client will be forced to
consider every directory in the system, even though it only cares about
a portion of them. There's also a bit of an information disclosure in
that any client could find the full list of directories, though that's
minor and in many ways has no serious implications.

`for_dir` also causes the initial sync of a new client to take much
longer, since it ends up processing a huge number of directories while
it actually knows about none at all.

The new mechanism has the client submit to the server the full list of
directories it has marked as clean, and the server determines which ones
have changed and returns that list to the client. This means that
neither party needs to waste time with directories the client has no
immediate interest in, and, for what little it's worth, that the client
cannot discover new directory ids.

There are some disadvantages though. The total bandwidth is always _at
least_ the same as `for_dir` in the single-root use-case, and the
majority of that is now upload instead of download bandwidth. The new
system also causes the client to hand a full list of interesting
directories to a potential MITM, though ultimately a MITM in the old
system could manipulate `for_dir` to cause the client to traverse all
interesting directories anyway.

It may be desirable to later submit the directory version and length in
a more compact form, e.g. a 32-bit hash, simply to reduce outgoing
bandwidth.