Fix warnings.
A couple were introduced last commit but missed in the deluge of other
warnings.

The rest are due to the new import resolution rules in Rust 1.15.0.
Apparently the new meaning of `use super::*` *also* includes the parent
module's imports, which is weird, but whatever. We suppress these
warnings instead of removing the imports since they are needed for
compatibility with the earlier Rust versions.